Refactored core and config package responsibilities to be cleaner.
