+++ spring-cloud-vault-config-databases/src/test/java/org/springframework/cloud/vault/config/databases/VaultConfigCassandraTests.java
@@ -15,8 +15,9 @@
  */
 package org.springframework.cloud.vault.config.databases;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.Assume.*;
+import com.datastax.driver.core.Cluster;
+import com.datastax.driver.core.PlainTextAuthProvider;
+import com.datastax.driver.core.Session;
 
 import java.net.InetSocketAddress;
 import java.sql.SQLException;
@@ -37,21 +38,20 @@ import org.springframework.cloud.vault.util.CanConnect;
 import org.springframework.cloud.vault.util.VaultRule;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-import com.datastax.driver.core.Cluster;
-import com.datastax.driver.core.PlainTextAuthProvider;
-import com.datastax.driver.core.Session;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.Assume.assumeTrue;
 
 /**
  * Integration tests using the cassandra secret backend. In case this test should fail because of SSL make sure you run
  * the test within the spring-cloud-vault-config/spring-cloud-vault-config directory as the keystore is referenced with
  * {@code ../work/keystore.jks}.
- * 
+ *
  * @author Mark Paluch
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = VaultConfigCassandraTests.TestApplication.class)
 @IntegrationTest({ "spring.cloud.vault.cassandra.enabled=true",
-		"spring.cloud.vault.cassandra.role=readonly" })
+"spring.cloud.vault.cassandra.role=readonly" })
 public class VaultConfigCassandraTests {
 
 	private final static String CASSANDRA_HOST = "localhost";
